Set compability mode when writing uncompressed LAS
Previously the compressed flag has been checked when setting the
compatibility mode. This caused later in the code, that no memory for
extra_bytes was allocated, and subsequently a segmentation fault error
when writing out points uncompressed AND in compatibility mode. It seems
that compatibility should work with uncompressed data as the example 7
of laszipdllexample.cpp suggests.